requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
attrs==23.2.0
disnake==2.9.1
frozenlist==1.4.1
greenlet==3.0.3
idna==3.6
multidict==6.0.4
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
python-dotenv==1.0.0
SQLAlchemy==2.0.0
typing_extensions==4.9.0
yarl==1.9.4
